requirements.txt
# flask==3.0.2
# pandas==2.2.1
# langchain==0.3.22
# langchain-huggingface==0.1.2
# faiss-cpu>=1.8.0
# sentence-transformers==2.6.0
# groq>=0.5.0  # Updated from 0.4.2
# python-dotenv==1.0.1
# langchain-community>=0.3.0 
flask==3.0.3
flask-cors==5.0.0
pandas==2.2.1
langchain==0.3.23
langchain-community==0.3.21
langchain-huggingface==0.1.2
faiss-cpu==1.8.0
sentence-transformers==2.6.0
groq==0.22.0
python-dotenv==1.0.1
httpx==0.27.2
httpcore==1.0.5
langdetect==1.0.9
PyPDF2==3.0.1
pytesseract==0.3.13
pillow==10.4.0
langsmith==0.1.125
py-googletrans==4.0.0
pdf2image==1.17.0
opencv-python==4.10.0.84 
google-api-python-client==2.146.0